refactor(cli): removes reusable CLI code to cli_base to enable multiple entrypoints
